pyproject.toml
[tool.poetry]
name = "week3-task1"
version = "0.1.0"
description = ""
authors = ["AntonDoncov <anton.dontsovgs@mail.ru>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pytest = "^7.4.0"
wemake-python-styleguide = "^0.18.0"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
fastapi = {version = "^0.103.1", extras = ["all"]}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
simplejson = "^3.19.1"


aiohttp = "^3.8.5"
asgi-lifespan = "^2.1.0"
aiokafka = "^0.8.1"
aiojobs = "^1.2.0"
fastapi-lifespan-manager = "^0.1.0"
pytest-asyncio = "^0.21.1"
Pillow = "^10.0.1"
pydantic = "^2.4.1"
pydantic-settings = "^2.0.3"
SQLAlchemy = "^2.0.21"
psycopg2-binary = "^2.9.8"
alembic = "^1.12.0"
opentracing = "^2.4.0"
prometheus-client = "^0.17.1"
jaeger-client = "^4.8.0"


[tool.poetry.dev-dependencies]
h5py = "^3"
scipy = "^1.11"
tensorflow-io = {extras = ["tensorflow"], version="^0.27"}
deepface = "^0.0.79"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
